+++ R/zzz.R
@@ -39,6 +39,7 @@ t61_env <- NULL
 .onAttach <- function(libname, pkgname) {
   op <- options()
   op.theme61 <- list(
+    theme61.auto_theme = TRUE,
     theme61.base_size = 10,
     theme61.default_save_format = "svg",
     theme61.open_e61_graph = FALSE,